requirements.txt
streamlit
langchain
google.generativeai
python-dotenv
pathlib
youtube_transcript_api